NFTImporter/requirements.txt
requests==2.25.1
pycryptodome==3.10.1